requirements.txt
# Web Scraping
scrapy>=2.11.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0
playwright>=1.40.0
selenium>=4.15.0

# Data Processing & Analysis
pandas>=2.0.0
openpyxl>=3.1.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Database
# SQLite3 viene incluido con Python

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# Optional: Claude API para parsing avanzado
anthropic>=0.18.0

# Logging
colorlog>=6.7.0
